process_monitoring_project/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
psutil==5.9.8
requests==2.32.3
pyinstaller==6.6.0
